pdf_to_pil_image/pyproject.toml
[tool.poetry]
name = "pdf_to_pil_image"
version = "0.1.0"
description = ""
authors = ["Marc Boissonneault <marc.boissonneault@elementai.com>"]

[[tool.poetry.source]]
name = "eai-core"
url = "https://pypi.elmt.io/repo/eai-core"
verify_ssl = true

[tool.poetry.dependencies]
python = "^3.7"
eai-image-loader = "^0.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
